Move tests to same package of the code being tested
