+++ app/src/test/java/nl/marc_apps/ovgo/utils/BitwiseOperationsUtilsTests.kt
@@ -1,6 +1,5 @@
-package nl.marc_apps.ovgo
+package nl.marc_apps.ovgo.utils
 
-import nl.marc_apps.ovgo.utils.BitwiseOperations
 import kotlin.test.*
 
 class BitwiseOperationsUtilsTests {
